perf: seed the org dashboard's role query so it stops SSR-ing nothing
useOrgRole reads `data?.membership.role ?? "LEARNER"` and
orgDetailsQueryKey(orgId) was prefetched NOWHERE in the repo. During SSR that
data is undefined, the role fails closed to LEARNER, every operator permission
check fails, and ~15 components under the org tree hit `if (!allowed) return
null` — so the org dashboard emitted no markup at all.

Two prefetches that already existed were dead on arrival behind that gate:
analytics/page.tsx seeds ["org-analytics", orgId] with a comment correctly
insisting the key must match — it does — and AnalyticsPageClient returned null
before ever reading it. Same for members/page.tsx. Seeding the role is what
makes those two pay off.

Worse than a blank shell: org home derives isOperator from that role, so the
server rendered the CONSUMER view and swapped to the operator dashboard after
hydration.

Splits the layout the way org-workspace already does — client component becomes
OrgDashboardShell, layout.tsx becomes a server wrapper. Checked first that the
org layout has zero existing duplications, so unlike the consultant layout this
rename cannot trip Sonar's new-code duplication gate.

Also fixes a latent bug found on the way. Organization has no `logo` column,
yet OrgDetailsResponse declared `logo: string | null` and the sidebar rendered
`org.organization.logo` at two call sites — so org avatars have never displayed,
silently. The route returns untyped NextResponse.json, so tsc was never asked to
compare the response against the type describing it; writing a typed server read
forced that comparison. Added `logo String?` (additive nullable, allowed under
the schema freeze) rather than deleting plumbing the UI clearly intends.

REQUIRES A DB PUSH before deploy — new nullable column.
